+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Fixed
 
+- [Config option `focus_on` didn't filter match list](https://github.com/fewieden/MMM-NFL/issues/55)
+
 ### Added
 
 ### Changed